Scripted adding predictions
