+++ etl/common.py
@@ -1,48 +1,43 @@
-#!/usr/bin/env python
+# environment variables
 from dotenv import load_dotenv
+import dotenv
 import os
-from sqlalchemy import create_engine, text
-from snowflake.sqlalchemy import URL
-import logging
-from snowflake.connector.pandas_tools import pd_writer
+# api
+import requests
 
-logging.basicConfig(
-    level=logging.INFO,
-    format="%(asctime)s %(levelname)s %(message)s",
-    datefmt="%Y-%m-%d %H:%M:%S",
-)
+# Load in env variables from .env file
+load_dotenv()
 
-account = os.getenv("SF_ACCOUNT_IDENTIFIER")
-username = os.getenv("SF_USERNAME")
-password = os.getenv("SF_PASSWORD")
-# database = "spotify"
-# warehouse = "computeWH"
-# schema = 'public'
-# role = 'accountadmin'
-# region = 'AWS_CA_CENTRAL_1'
+client_id = os.getenv("CLIENT_ID")
+client_secret = os.getenv("CLIENT_SECRET")
+access_token = os.getenv("ACCESS_TOKEN")
+refresh_token = os.getenv("REFRESH_TOKEN")
+expires_at = os.getenv("EXPIRES_AT") # TODO: Implement expiry logic for refresh token
 
-def get_engine(database: str, schema: str):
-    connection_parameters = {
-        'account': account,
-        'database': database,
-        # 'warehouse': warehouse,
-        'schema': schema,
-        'user': username,
-        'password': password,
-        # 'role': role,
-        # 'region': region
-    }    
-    try:
-        url = URL(**connection_parameters)
-        engine = create_engine(url)
-        logging.info(f"Creating engine: {engine}")
-        return engine
-    except Exception as e:
-        logging.warning(f"Error occurred: {str(e)}")
 
-def write_to_sf(engine, tbl_name, dataframe, if_exists="replace"):
-    with engine.connect() as conn:
-        logging.info(f"Writing to {tbl_name}...")
-        dataframe.columns = map(str.upper, dataframe.columns)
-        dataframe.to_sql(name=tbl_name, con=conn, if_exists=if_exists, index=False, method=pd_writer)
-    
+# Returns authorization header
+def get_auth_header(token):
+    return {"Authorization": f"Bearer {token}"}
+
+# Returns API token
+def get_token():
+    url = "https://accounts.spotify.com/api/token"
+    request_body = {
+        "grant_type": "refresh_token",
+        "refresh_token": refresh_token,
+        "client_id": client_id,
+        "client_secret": client_secret,
+    }
+
+    response = requests.post(url, data=request_body)
+    new_token_info = response.json()
+
+    dotenv_file = dotenv.find_dotenv()
+    dotenv.load_dotenv(dotenv_file)
+
+    os.environ["ACCESS_TOKEN"] = new_token_info['access_token']
+
+    # Write changes to .env file.
+    dotenv.set_key(dotenv_file, "ACCESS_TOKEN", os.environ["ACCESS_TOKEN"])
+
+    return new_token_info['access_token']
